breaking shared method out into a utils file
